fix: refactoriza la lógica de búsqueda de productos para mejorar el manejo de preferencias y la paginación.
