+++ src/main/java/com/rescuebites/api/shared/services/implementations/PublicSearchServiceImpl.java
@@ -3,8 +3,8 @@ package com.rescuebites.api.shared.services.implementations;
 import com.rescuebites.api.shared.controllers.responses.SearchProductResponse;
 import com.rescuebites.api.shared.controllers.responses.SearchSuggestion;
 import com.rescuebites.api.shared.services.interfaces.IPublicSearchService;
-import com.rescuebites.api.shared.services.search.ProductSearchStrategy;
-import com.rescuebites.api.shared.services.search.SuggestionSearchStrategy;
+import com.rescuebites.api.shared.services.strategy.ProductSearchStrategy;
+import com.rescuebites.api.shared.services.strategy.SuggestionSearchStrategy;
 import com.rescuebites.api.shared.utils.SearchUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
