+++ src/main/java/com/rescuebites/api/shared/utils/Constants.java
@@ -7,7 +7,5 @@ public class Constants {
     private Constants() {
     }
 
-    public static final PageRequest PRE_FILTER_PAGE_FOR_SEARCH = PageRequest.of(0, 100);
-
     public static final PageRequest PRE_FILTER_PAGE_FOR_SUGGESTIONS = PageRequest.of(0, 5);
 }
